requirements.txt
# Core dependencies
fastmcp>=0.1.0
mcp>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
openai>=1.0.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
httpx>=0.24.0

# Development dependencies
black>=23.0.0
pylint>=2.17.0
mypy>=1.0.0